# backend/api/controllers/backend_api/openapi/openapi.yml
openapi: 3.1.0
info:
  title: backend_api
  version: 0.1.0
servers:
  - url: http://localhost:8888/backend-api
paths:
  /ddsp/train:
    post:
      summary: Train Ddsp
      operationId: TrainDdsp
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_train_ddsp_ddsp_train_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Features'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /ddsp/generate:
    post:
      summary: Generate Audio From Ddsp
      operationId: GenerateAudioFromDdsp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DDSPGenerateParams'
      responses:
        '200':
          description: Successful Response
          content:
            audio/wav:
              schema:
                type: string
                format: binary
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /diffusion/train:
    post:
      summary: Train Diffusion
      description: WAVとMIDIをアラインし、DDSPガイダンス付きDiffusionで各楽器の合成パラメータを生成する。
      operationId: TrainDiffusion
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_train_diffusion_diffusion_train_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Features'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /diffusion/generate:
    post:
      summary: Generate Params From Diffusion
      description: 音符列と楽器IDから合成パラメータを生成（Diffusionモデルを使用）
      operationId: GenerateParamsFromDiffusion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiffusionGenerateParams'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DDSPGenerateParams'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /fluidsynth/train:
    post:
      summary: Train Fluidsynth
      description: WAVとMIDIをアラインし、各楽器のアライン済みMIDIから FluidSynth で音源を生成してZIPで返す。
      operationId: TrainFluidsynth
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_train_fluidsynth_fluidsynth_train_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /fluidsynth/generate:
    post:
      summary: Generate Audio From Fluidsynth
      description: 音符列から MIDI を作成し、FluidSynth で音源を合成する。
      operationId: GenerateAudioFromFluidsynth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FluidsynthGenerateParams'
      responses:
        '200':
          description: Successful Response
          content:
            audio/wav:
              schema:
                type: string
                format: binary
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    Note:
      type: object
      required:
        - start
        - frequency
        - duration
      properties:
        start:
          type: number
          title: Start
        frequency:
          type: number
          title: Frequency
        duration:
          type: number
          title: Duration
      title: Note
    Feature:
      type: object
      required:
        - instrument_name
        - pitch
        - loudness
        - z_feature
        - notes
      properties:
        instrument_name:
          type: string
          title: Instrument Name
        pitch:
          type: array
          items:
            type: number
          title: Pitch
        loudness:
          type: array
          items:
            type: number
          title: Loudness
        z_feature:
          type: array
          items:
            type: array
            items:
              type: number
          title: Z Feature
        notes:
          type: array
          items:
            $ref: '#/components/schemas/Note'
          title: Notes
      title: Feature
    Features:
      type: object
      required:
        - features
      properties:
        features:
          type: array
          items:
            $ref: '#/components/schemas/Feature'
          title: Features
      title: Features
    DDSPGenerateParams:
      type: object
      required:
        - pitch
        - loudness
        - z_feature
      properties:
        pitch:
          type: array
          items:
            type: number
          title: Pitch
        loudness:
          type: array
          items:
            type: number
          title: Loudness
        z_feature:
          type: array
          items:
            type: array
            items:
              type: number
          title: Z Feature
      title: DDSPGenerateParams
    DiffusionGenerateParams:
      type: object
      required:
        - notes
        - instrument_name
        - signal_length
      properties:
        notes:
          type: array
          items:
            $ref: '#/components/schemas/Note'
          title: Notes
        instrument_name:
          type: string
          title: Instrument Name
        signal_length:
          type: integer
          title: Signal Length
      title: DiffusionGenerateParams
    FluidsynthGenerateParams:
      type: object
      required:
        - notes
        - instrument_name
        - signal_length
      properties:
        notes:
          type: array
          items:
            $ref: '#/components/schemas/Note'
          title: Notes
        instrument_name:
          type: string
          title: Instrument Name
        signal_length:
          type: integer
          title: Signal Length
      title: FluidsynthGenerateParams
    Body_train_ddsp_ddsp_train_post:
      type: object
      required:
        - wav_file
        - midi_file
      properties:
        wav_file:
          type: string
          format: binary
          title: Wav File
          description: WAVファイル
        midi_file:
          type: string
          format: binary
          title: Midi File
          description: MIDIファイル
      title: Body_train_ddsp_ddsp_train_post
    Body_train_diffusion_diffusion_train_post:
      type: object
      required:
        - wav_file
        - midi_file
      properties:
        wav_file:
          type: string
          format: binary
          title: Wav File
          description: WAVファイル
        midi_file:
          type: string
          format: binary
          title: Midi File
          description: MIDIファイル
      title: Body_train_diffusion_diffusion_train_post
    Body_train_fluidsynth_fluidsynth_train_post:
      type: object
      required:
        - wav_file
        - midi_file
      properties:
        wav_file:
          type: string
          format: binary
          title: Wav File
          description: WAVファイル
        midi_file:
          type: string
          format: binary
          title: Midi File
          description: MIDIファイル
      title: Body_train_fluidsynth_fluidsynth_train_post
    HTTPValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
      title: HTTPValidationError
    ValidationError:
      type: object
      required:
        - loc
        - msg
        - type
      properties:
        loc:
          type: array
          items:
            anyOf:
              - type: string
              - type: integer
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      title: ValidationError